% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-handle.R
\name{check_dimension}
\alias{check_dimension}
\title{Checks the dimensions of input objects for perturbCS}
\usage{
check_dimension(data, psi_fn, mom_constr, par_constr)
}
\arguments{
\item{data}{A data matrix.}

\item{psi_fn}{A function.}

\item{mom_constr}{A list with two functions.}

\item{par_constr}{A list with three matrices where first and third matrices
must have two rows.}
}
\value{
Returns an error message if dimensions of inputs do not align.
}
\description{
Checks the dimensions of input objects for perturbCS
}
