% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-handle.R
\name{check_type}
\alias{check_type}
\title{Checks classes of input objects for perturbCS}
\usage{
check_type(
  data,
  psi_fn,
  mom_constr,
  par_constr,
  max_perturb,
  alpha,
  tol,
  boot,
  method
)
}
\arguments{
\item{data}{A data matrix.}

\item{psi_fn}{A function.}

\item{mom_constr}{A list with two functions.}

\item{par_constr}{A list with three matrices where first and third matrices
must have two rows.}

\item{max_perturb}{A list with two vectors where the second vector must be
nonnegative.}

\item{alpha}{A numeric vector with element(s) between 0 and 1 (exclusive).}

\item{tol}{A nonnegative number.}

\item{boot}{A positive integer.}

\item{method}{A string.}
}
\value{
Returns an error message is an input object is of invalid class.
}
\description{
Checks classes of input objects for perturbCS
}
