% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-Rmosek.R
\name{gen_perturb_Rmosek}
\alias{gen_perturb_Rmosek}
\title{Create a list of perturbation draws tailored to Rmosek solver}
\usage{
gen_perturb_Rmosek(data, mom_constr, pconstr, max_perturb)
}
\arguments{
\item{data}{A data matrix.}

\item{mom_constr}{A list with two functions.}

\item{pconstr}{A list with three matrices.}

\item{max_perturb}{A list with two vectors.}
}
\value{
Returns perturbation draws for functional of interest (objective
function of linear programs) and moment constraints (feasible region).
}
\description{
Create a list of perturbation draws tailored to Rmosek solver
}
