% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-lpSolve.R
\name{mom_constr_lpSolve}
\alias{mom_constr_lpSolve}
\title{Create a list defining the moment constraints tailored to lpSolve solver}
\usage{
mom_constr_lpSolve(data, mom_constr)
}
\arguments{
\item{data}{A data matrix.}

\item{mom_constr}{A list with two functions.}
}
\value{
Returns linear constraint matrix and  constraint upper and lower
bounds defined by moment inequalities reshaped for use in lpSolve.
}
\description{
Create a list defining the moment constraints tailored to lpSolve solver
}
