% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-Rmosek.R
\name{mom_constr_Rmosek}
\alias{mom_constr_Rmosek}
\title{Create a list defining the moment constraints tailored to Rmosek solver}
\usage{
mom_constr_Rmosek(data, mom_constr)
}
\arguments{
\item{data}{A data matrix.}

\item{mom_constr}{A list with two functions.}
}
\value{
Returns linear constraint matrix and  constraint upper and lower
bounds defined by moment inequalities reshaped for use in Rmosek.
}
\description{
Create a list defining the moment constraints tailored to Rmosek solver
}
