% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-lpSolve.R
\name{par_constr_lpSolve}
\alias{par_constr_lpSolve}
\title{Create a list defining the parameter space tailored to lpSolve solver}
\usage{
par_constr_lpSolve(data, mom_constr, par_constr)
}
\arguments{
\item{data}{A data matrix.}

\item{mom_constr}{A list with two functions.}

\item{par_constr}{A list with three matrices.}
}
\value{
Returns linear constraint matrix, constraint upper and lower bounds
and parameter space constraints reshaped for use in lpSolve.
}
\description{
Create a list defining the parameter space tailored to lpSolve solver
}
