% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-Rmosek.R
\name{par_constr_Rmosek}
\alias{par_constr_Rmosek}
\title{Create a list defining the parameter space tailored to Rmosek solver}
\usage{
par_constr_Rmosek(data, mom_constr, par_constr)
}
\arguments{
\item{data}{A data matrix.}

\item{mom_constr}{A list with two functions.}

\item{par_constr}{A list with three matrices.}
}
\value{
Returns linear constraint matrix, constraint upper and lower bounds
and parameter space constraints reshaped for use in Rmosek.
}
\description{
Create a list defining the parameter space tailored to Rmosek solver
}
