% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbCS.R
\name{perturbCS}
\alias{perturbCS}
\title{Construct confidence set for functional of partially identified
parameter}
\usage{
perturbCS(
  data,
  psi_fn,
  mom_constr,
  par_constr = NULL,
  max_perturb = list(psi = c(-1e-06, 1e-06), constr = c(0, 1e-06)),
  alpha = c(0.1, 0.05, 0.01),
  tol = 1,
  boot = 999,
  method = "lpSolve"
)
}
\arguments{
\item{data}{A data matrix.}

\item{psi_fn}{A function that returns a numeric vector of intercept and
coefficients of the linear functional.}

\item{mom_constr}{A list with two elements, first element is a function that
returns the matrix of coefficients for linear moment constraints and the
second element is a function that returns the lower and upper bounds of
the constraints as a matrix with two rows; if no lower or upper bound
for a moment constraint, set \code{-Inf} or \code{Inf} for the
appropriate element of the bound matrix.}

\item{par_constr}{A list with three elements, first element is a matrix with
two rows for the lower and upper bounds on the constraints for individual
coordinates of the parameter space, second element is a matrix with
coefficients that define the planes that restrict the parameter space,
third element is a matrix with two rows for the lower and upper bounds
of the linear constraints defined by the planes; if no lower or upper
bound for a moment constraint, set \code{-Inf} or \code{Inf} for the
appropriate element of the bound matrix.}

\item{max_perturb}{A list with two elements, first element is a numeric
vector with two values corresponding to the lower and upper bounds for
perturbations to the functional of interest, second element is a numeric
vector with two values corresponding to the lower and upper bounds for
perturbations to the linear moment constraints; default set to
\code{max_perturb = list("psi"=c(-1e-6,1e-6),"constr"=c(0,1e-6))}.}

\item{alpha}{A numeric vector of values for desired sizes of the confidence
set; default set to \code{alpha = c(.10,.05,.01)}.}

\item{tol}{A numeric value for tolerance level to threshold quantile
selection of bootstrap critical values in small samples; default set to
\code{tol = 1}.}

\item{boot}{A numeric value for the bootstrap sample size; default set to
\code{boot = 999}.}

\item{method}{A string for the solver to be used; currently only
\code{method = "lpSolve"} and \code{"Rmosek"} are supported; default set
to \code{method = "lpSolve"}.}
}
\value{
A bootstrap confidence set for functional of interest with specified
size. Additionally, estimated lower and upper bounds of the identified
set are reported. Logical output for whether the finite-sample threshold
is used and the number of valid bootstrap samples with nonempty feasible
regions for the perturbed bootstrap linear programs are also reported.
}
\description{
Construct a confidence set for linear functional of partially identified
parameter defined by linear moment inequalities. The confidence set is
constructed by bootstrapping perturbed linear programs that define the lower
and upper bounds of the identified set for the functional of interest.
}
