% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-lpSolve.R
\name{perturbLP_lpSolve}
\alias{perturbLP_lpSolve}
\title{Solve perturbed linear programs using lpSolve solver}
\usage{
perturbLP_lpSolve(data, psi_fn, mom_constr, pconstr, perturb)
}
\arguments{
\item{data}{A data matrix.}

\item{psi_fn}{A function.}

\item{mom_constr}{A list with two functions.}

\item{pconstr}{A list with three matrices.}

\item{perturb}{A list with a vector and two matrices.}
}
\value{
Returns values of four perturbed linear programs for a given
perturbation draw \code{perturb}. If feasible region empty, a vector
of \code{NA}s is returned.
}
\description{
Solve perturbed linear programs using lpSolve solver
}
